Auth check controller

+++ app/Http/Controllers/FollowController.php
@@ -12,72 +12,87 @@ class FollowController extends Controller
 {
     public function getFollowing(Request $request)
     {
-        $id = $request->id;
-        //var_dump($id);
-        $user = User::findOrFail($id);
-
-        $following = User::find($user->id)->follows;
-        $data = ['users' => []];
-
-        foreach ($following as $follow)
+        if(Auth::check())
         {
-            $user_id = $follow->user_followed;
+            $id = $request->id;
+            //var_dump($id);
+            $user = User::findOrFail($id);
 
-            $user = User::find($user_id);
-            array_push($data['users'], $user);
+            $following = User::find($user->id)->follows;
+            $data = ['users' => []];
+
+            foreach ($following as $follow)
+            {
+                $user_id = $follow->user_followed;
+
+                $user = User::find($user_id);
+                array_push($data['users'], $user);
+            }
+
+            $response = ['status' => 'success','data' => $data];
+            return json_encode($response);
         }
 
-        $response = ['status' => 'success','data' => $data];
-        return json_encode($response);
 
     }
 
     public function getFollowers(Request $request)
     {
-        $id = $request->id;
-        $user = User::findOrFail($id);
-
-        $following = User::find($user->id)->followed;
-        $data = ['users' => []];
-
-        foreach ($following as $follow)
+        if(Auth::user())
         {
-            $user_id = $follow->user_following;
+            $id = $request->id;
+            $user = User::findOrFail($id);
 
-            $user = User::find($user_id);
-            array_push($data['users'], $user);
+            $following = User::find($user->id)->followed;
+            $data = ['users' => []];
+
+            foreach ($following as $follow)
+            {
+                $user_id = $follow->user_following;
+
+                $user = User::find($user_id);
+                array_push($data['users'], $user);
+            }
+
+            $response = ['status' => 'success','data' => $data];
+            return json_encode($response);
         }
-
-        $response = ['status' => 'success','data' => $data];
-        return json_encode($response);
 
     }
 
     public function unfollow(Request $request)
     {
-        $id_user = Auth::id();
-        $id_following = $request->id_following;
+        if(Auth::user())
+        {
+            $id_user = Auth::id();
+            $id_following = $request->id_following;
 
-        $follow = Follow::where('user_following',$id_user)->where('user_followed',$id_following);
+            $follow = Follow::where('user_following',$id_user)->where('user_followed',$id_following);
 
-        $follow->delete();
+            $follow->delete();
 
-        return ['status' => 'success'];
+            return ['status' => 'success'];
+        }
+
     }
 
     public function follow(Request $request)
     {
-        $id_user = Auth::id();
-        $id_following = $request->id_following;
+        if(Auth::user())
+        {
+            $id_user = Auth::id();
+            $id_following = $request->id_following;
 
-        $follow = new Follow();
+            $follow = new Follow();
 
-        $follow->user_following = $id_user;
-        $follow->user_followed = $id_following;
+            $follow->user_following = $id_user;
+            $follow->user_followed = $id_following;
 
-        $follow->save();
+            $follow->save();
 
-        return ['status' => 'success'];
+            return ['status' => 'success'];
+        }
+
     }
 
 }
